backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["marioyordanoff <mario@almalabs.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = "^0.30.4"
langchain = "^0.2.11"
qdrant-client = "^1.10.1"
python-decouple = "^3.8"
openai = "^1.37.1"
langchain-community = "^0.2.10"
beautifulsoup4 = "^4.12.3"
tiktoken = "^0.7.0"
websockets = "^12.0"
modal = "^0.63.98"
langchain-core = "^0.2.26"
langchain-openai = "^0.1.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
